+++ components/blog/Tag.js
@@ -1,7 +1,25 @@
-import styles from '@/styles/Blog.module.scss';
+import Link from 'next/link';
+import styled from 'styled-components';
 
-const Tag = ({ text }) => {
-  return <div className={styles.tag}>{text}</div>;
+export const TagContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  max-height: 12rem;
+  overflow: hidden;
+  gap: 0.5rem;
+  margin-bottom: -1rem;
+  a {
+    text-decoration: none;
+  }
+`;
+
+const TagLink = styled(Link)`
+  background-color: #8cd5e8;
+  padding: 0.5rem 1rem;
+  border-radius: 2rem;
+  font-weight: bold;
+`;
+
+export const Tag = ({ text }) => {
+  return <TagLink href={`/blog/category/${text}`}>{text}</TagLink>;
 };
-
-export default Tag;